fix(fe-fpm-writer): review comments
